Alert decimal to float

+++ CloudBees.BLL/DTOs/AlertDTO.cs
@@ -11,7 +11,7 @@ public class AlertDTO
 {
     public string Id { get; set; }
     public string Type { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal Longitude { get; set; }
+    public float Latitude { get; set; }
+    public float Longitude { get; set; }
     public string Location { get; set; }
 }
